Locate32 -> 2.99.6.7290

+++ Locate/Locate32/LocateApp.h
@@ -505,7 +505,7 @@ inline BOOL CLocateApp::IsUpdating() const
 
 inline void CLocateApp::ChangeAndAlloc(LPWSTR& pVar,LPCWSTR szText)
 {
-	SIZE_T dwLength=istrlenw(szText);
+	DWORD dwLength=istrlenw(szText);
 	if (pVar!=NULL)
 		delete[] pVar;
 	pVar=new WCHAR [dwLength+1];
